Contains all casting within CSR implemenation

+++ src/main/scala/leapfrogTriejoin/TrieIterator.scala
@@ -6,6 +6,6 @@ trait TrieIterator extends LinearIterator {
 
   def up(): Unit // O(log N)
 
-  def translate(keys: Array[Int]): Array[Long]
+  def translate(keys: Array[Long]): Array[Long]
 
 }
